+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/SequenceFeatureDetails.tsx
@@ -3,11 +3,11 @@ import { Suspense, lazy, useEffect, useRef, useState } from 'react'
 import { Button, Typography } from '@mui/material'
 import { observer } from 'mobx-react'
 
-import { useFeatureSequence } from './useFeatureSequence'
 import { ErrorMessage, LoadingEllipses } from '../../ui'
 import SequenceFeatureMenu from './dialogs/SequenceFeatureMenu'
 import SequenceTypeSelector from './dialogs/SequenceTypeSelector'
 import { SimpleFeature, getSession } from '../../util'
+import { useFeatureSequence } from '../../util/useFeatureSequence'
 
 import type { SimpleFeatureSerialized } from '../../util'
 import type { BaseFeatureWidgetModel } from '../stateModelFactory'
