+++ plugins/svg/src/SvgFeatureRenderer/components/Segments.tsx
@@ -32,9 +32,7 @@ const Segments = observer(function Segments(props: {
   const theme = useTheme()
   const c = readConfObject(config, 'color2', { feature })
   const color2 = c === '#f0f' ? stripAlpha(theme.palette.text.secondary) : c
-
   const { left = 0, top = 0, width = 0, height = 0 } = featureLayout.absolute
-
   const y = top + height / 2
   return (
     <>
